ia_backend/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
python-dotenv
python-multipart